rename: icon 컴포넌트들의 파일명을 원래대로 수정 (#101)
