+++ src/components/Icon/index.js
@@ -1,7 +1,7 @@
-import IconHome from './iconHome2';
-import IconProfile from './iconProfile2';
-import IconTimetable from './iconTimetable2';
-import IconNotice from './iconNotice2';
-import IconEmployee from './iconEmployee2';
+import IconHome from './iconHome';
+import IconProfile from './iconProfile';
+import IconTimetable from './iconTimetable';
+import IconNotice from './iconNotice';
+import IconEmployee from './iconEmployee';
 
 export { IconHome, IconProfile, IconTimetable, IconNotice, IconEmployee };
